test: remove integration tests in favor of unit
