+++ quic/s2n-quic-core/src/frame/tests.rs
@@ -1,11 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::frame::FrameRef;
 use bolero::check;
 use s2n_codec::{assert_codec_round_trip_bytes_mut, Encoder, EncoderLenEstimator, EncoderValue};
-use s2n_quic_core::frame::FrameRef;
 
-fn main() {
+#[test]
+fn round_trip() {
     check!().for_each(|input| {
         let mut input = input.to_vec();
         let frames = assert_codec_round_trip_bytes_mut!(FrameRef, &mut input);
