+++ quic/s2n-quic-core/src/packet/number/mod.rs
@@ -62,6 +62,9 @@ pub mod map;
 #[cfg(feature = "alloc")]
 pub use map::Map;
 
+#[cfg(test)]
+mod tests;
+
 //= https://www.rfc-editor.org/rfc/rfc9000#section-17.1
 //# the sender MUST use a packet number size able to represent more than
 //# twice as large a range as the difference between the largest
@@ -224,144 +227,4 @@ fn decode_packet_number(
     };
 
     PacketNumber::from_varint(candidate_pn, space)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use bolero::check;
-
-    fn new(value: VarInt) -> PacketNumber {
-        PacketNumberSpace::Initial.new_packet_number(value)
-    }
-
-    /// This implementation tries to closely follow the RFC pseudo code so it's
-    /// easier to ensure it matches.
-    #[allow(clippy::blocks_in_if_conditions)]
-    fn rfc_decoder(largest_pn: u64, truncated_pn: u64, pn_nbits: usize) -> u64 {
-        use std::panic::catch_unwind as catch;
-
-        //= https://www.rfc-editor.org/rfc/rfc9000#section-A.3
-        //= type=test
-        //# DecodePacketNumber(largest_pn, truncated_pn, pn_nbits):
-        //#   expected_pn  = largest_pn + 1
-        //#   pn_win       = 1 << pn_nbits
-        //#   pn_hwin      = pn_win / 2
-        //#   pn_mask      = pn_win - 1
-        //#   // The incoming packet number should be greater than
-        //#   // expected_pn - pn_hwin and less than or equal to
-        //#   // expected_pn + pn_hwin
-        //#   //
-        //#   // This means we cannot just strip the trailing bits from
-        //#   // expected_pn and add the truncated_pn because that might
-        //#   // yield a value outside the window.
-        //#   //
-        //#   // The following code calculates a candidate value and
-        //#   // makes sure it's within the packet number window.
-        //#   // Note the extra checks to prevent overflow and underflow.
-        //#   candidate_pn = (expected_pn & ~pn_mask) | truncated_pn
-        //#   if candidate_pn <= expected_pn - pn_hwin and
-        //#     candidate_pn < (1 << 62) - pn_win:
-        //#     return candidate_pn + pn_win
-        //#   if candidate_pn > expected_pn + pn_hwin and
-        //#     candidate_pn >= pn_win:
-        //#     return candidate_pn - pn_win
-        //#   return candidate_pn
-        let expected_pn = largest_pn + 1;
-        let pn_win = 1 << pn_nbits;
-        let pn_hwin = pn_win / 2;
-        let pn_mask = pn_win - 1;
-
-        let candidate_pn = (expected_pn & !pn_mask) | truncated_pn;
-        if catch(|| {
-            candidate_pn <= expected_pn.checked_sub(pn_hwin).unwrap()
-                && candidate_pn < (1u64 << 62).checked_sub(pn_win).unwrap()
-        })
-        .unwrap_or_default()
-        {
-            return candidate_pn + pn_win;
-        }
-
-        if catch(|| {
-            candidate_pn > expected_pn.checked_add(pn_hwin).unwrap() && candidate_pn >= pn_win
-        })
-        .unwrap_or_default()
-        {
-            return candidate_pn - pn_win;
-        }
-
-        candidate_pn
-    }
-
-    #[test]
-    fn truncate_expand_test() {
-        check!()
-            .with_type()
-            .cloned()
-            .for_each(|(largest_pn, expected_pn)| {
-                let largest_pn = new(largest_pn);
-                let expected_pn = new(expected_pn);
-                if let Some(truncated_pn) = expected_pn.truncate(largest_pn) {
-                    assert_eq!(expected_pn, truncated_pn.expand(largest_pn));
-                }
-            });
-    }
-
-    #[test]
-    fn rfc_differential_test() {
-        check!()
-            .with_type()
-            .cloned()
-            .for_each(|(largest_pn, truncated_pn)| {
-                let largest_pn = new(largest_pn);
-                let space = largest_pn.space();
-                let truncated_pn = TruncatedPacketNumber {
-                    space,
-                    value: truncated_pn,
-                };
-                let rfc_value = rfc_decoder(
-                    largest_pn.as_u64(),
-                    truncated_pn.into_u64(),
-                    truncated_pn.bitsize(),
-                )
-                .min(VarInt::MAX.as_u64());
-                let actual_value = truncated_pn.expand(largest_pn).as_u64();
-
-                assert_eq!(
-                    actual_value,
-                    rfc_value,
-                    "diff: {}",
-                    actual_value
-                        .checked_sub(rfc_value)
-                        .unwrap_or(rfc_value - actual_value)
-                );
-            });
-    }
-
-    #[test]
-    fn example_test() {
-        macro_rules! example {
-            ($largest:expr, $truncated:expr, $expected:expr) => {{
-                let largest = new(VarInt::from_u32($largest));
-                let truncated = TruncatedPacketNumber::new($truncated, PacketNumberSpace::Initial);
-                let expected = new(VarInt::from_u32($expected));
-                assert_eq!(truncated.expand(largest), expected);
-            }};
-        }
-
-        example!(0xa82e1b31, 0x9b32u16, 0xa82e9b32);
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // snapshot tests don't work on miri
-    fn size_of_snapshots() {
-        use core::mem::size_of;
-        use insta::assert_debug_snapshot;
-
-        assert_debug_snapshot!("PacketNumber", size_of::<PacketNumber>());
-        assert_debug_snapshot!("PacketNumberLen", size_of::<PacketNumberLen>());
-        assert_debug_snapshot!("PacketNumberSpace", size_of::<PacketNumberSpace>());
-        assert_debug_snapshot!("ProtectedPacketNumber", size_of::<ProtectedPacketNumber>());
-        assert_debug_snapshot!("TruncatedPacketNumber", size_of::<TruncatedPacketNumber>());
-    }
 }
