+++ quic/s2n-quic-core/src/transport/parameters/tests.rs
@@ -1,11 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::transport::parameters::{ClientTransportParameters, ServerTransportParameters};
 use bolero::check;
 use s2n_codec::assert_codec_round_trip_bytes;
-use s2n_quic_core::transport::parameters::{ClientTransportParameters, ServerTransportParameters};
 
-fn main() {
+#[test]
+fn round_trip() {
     check!().for_each(|input| {
         if input.is_empty() {
             return;
