+++ quic/s2n-quic-crypto/Cargo.toml
@@ -31,8 +31,3 @@ hex-literal = "0.3"
 insta = { version = "1", features = ["json"] }
 pretty-hex = "0.3"
 s2n-quic-core = { path = "../s2n-quic-core", features = ["testing"] }
-
-[[test]]
-name = "fuzz_target"
-path = "tests/fuzz_target/main.rs"
-harness = false
